refactor(process): move Monitor and Http classes to app\process namespace
- Move Monitor.php from process/ to app/process/
- Create new Http.php file in app/process/ extending Webman\App
- Update namespace for both Monitor and Http classes to app\process
- Modify Monitor class to use $loadedFiles array for tracking included files
- Update file paths and reload logic in Monitor class

+++ app/process/Monitor.php
@@ -12,7 +12,7 @@
  * @license   http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace process;
+namespace app\process;
 
 use FilesystemIterator;
 use RecursiveDirectoryIterator;
@@ -36,6 +36,11 @@ class Monitor
      * @var array
      */
     protected $extensions = [];
+
+    /**
+     * @var array
+     */
+    protected $loadedFiles = [];
 
     /**
      * @var string
@@ -84,6 +89,12 @@ class Monitor
         static::resume();
         $this->paths = (array)$monitorDir;
         $this->extensions = $monitorExtensions;
+        foreach (get_included_files() as $index => $file) {
+            $this->loadedFiles[$file] = $index;
+            if (strpos($file, 'webman-framework/src/support/App.php')) {
+                break;
+            }
+        }
         if (!Worker::getAllWorkers()) {
             return;
         }
@@ -134,13 +145,17 @@ class Monitor
             }
             // check mtime
             if (in_array($file->getExtension(), $this->extensions, true) && $lastMtime < $file->getMTime()) {
+                $lastMtime = $file->getMTime();
+                if (DIRECTORY_SEPARATOR === '/' && isset($this->loadedFiles[$file->getRealPath()])) {
+                    echo "$file updated but cannot be reloaded because only auto-loaded files support reload.\n";
+                    continue;
+                }
                 $var = 0;
                 exec('"'.PHP_BINARY . '" -l ' . $file, $out, $var);
-                $lastMtime = $file->getMTime();
                 if ($var) {
                     continue;
                 }
-                echo $file . " update and reload\n";
+                echo $file . " updated and reload\n";
                 // send SIGUSR1 signal to master process for reload
                 if (DIRECTORY_SEPARATOR === '/') {
                     posix_kill(posix_getppid(), SIGUSR1);
